refactor publish subscribe features
