+++ src/NServiceBus.Core.Tests/Config/When_no_custom_configuration_source_is_specified.cs
@@ -15,7 +15,6 @@ namespace NServiceBus.Core.Tests.Config
 
             config.SendOnly();
             config.TypesToScanInternal(new[] { typeof(ConfigSectionValidatorFeature) });
-            config.DisableFeature<MessageDrivenSubscriptions>();
             config.EnableFeature<ConfigSectionValidatorFeature>();
 
             var endpoint = await Endpoint.Start(config);
