+++ src/NServiceBus.Core.Tests/Config/When_users_override_the_configuration_source.cs
@@ -16,7 +16,6 @@ namespace NServiceBus.Core.Tests.Config
 
             builder.SendOnly();
             builder.TypesToScanInternal(new[] { typeof(ConfigSectionValidatorFeature) });
-            builder.DisableFeature<MessageDrivenSubscriptions>();
             builder.EnableFeature<ConfigSectionValidatorFeature>();
             builder.CustomConfigurationSource(new UserConfigurationSource());
