+++ src/NServiceBus.Core.Tests/NServiceBus.Core.Tests.csproj
@@ -183,8 +183,7 @@
     <Compile Include="Routing\EndpointInstancesTests.cs" />
     <Compile Include="Transports\MSMQ\InstanceMapping\InstanceMappingFileParserTests.cs" />
     <Compile Include="Transports\MSMQ\InstanceMapping\InstanceMappingFileMonitorTests.cs" />
-    <Compile Include="Routing\MessageDrivenSubscribeTerminatorTests.cs" />
-    <Compile Include="Routing\MessageDrivenUnsubscribeTerminatorTests.cs" />
+    <Compile Include="Routing\MessageDrivenSubscriptionManagerTests.cs" />
     <Compile Include="Routing\Routers\MulticastPublishRouterBehaviorTests.cs" />
     <Compile Include="Routing\Routers\UnicastReplyRouterConnectorTests.cs" />
     <Compile Include="Routing\Routers\UnicastPublishRouterConnectorTests.cs" />
