+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -157,10 +157,16 @@
     <Compile Include="Routing\MessageDrivenSubscriptions\ConfiguredPublishers.cs" />
     <Compile Include="Routing\ConfiguredUnicastRoutes.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\IPublisherSource.cs" />
+    <Compile Include="Routing\MessageDrivenSubscriptions\MessageDrivenPublishSubscribeProvider.cs" />
+    <Compile Include="Routing\MessageDrivenSubscriptions\MessageDrivenSubscriptionManager.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\NamespacePublisherSource.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\TypePublisherSource.cs" />
-    <Compile Include="Routing\NativePublishSubscribeFeature.cs" />
+    <Compile Include="Routing\NativePublishSubscribeProvider.cs" />
     <Compile Include="Routing\ReplyAddressFeature.cs" />
+    <Compile Include="Routing\Routers\IPublishRouter.cs" />
+    <Compile Include="Routing\IPublishSubscribeProvider.cs" />
+    <Compile Include="Routing\Routers\PublishRouterBehavior.cs" />
+    <Compile Include="Routing\PublishSubscribeFeature.cs" />
     <Compile Include="Routing\RoutingComponent.cs" />
     <Compile Include="Routing\TypeRouteSource.cs" />
     <Compile Include="Routing\IDistributionPolicy.cs" />
@@ -410,9 +416,6 @@
     <Compile Include="Routing\Routers\UnicastSendRouterConnector.cs" />
     <Compile Include="Routing\UnicastAddressTag.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\IInitializableSubscriptionStorage.cs" />
-    <Compile Include="Routing\MessageDrivenSubscriptions\MessageDrivenUnsubscribeTerminator.cs" />
-    <Compile Include="Routing\MessageDrivenSubscriptions\MessageDrivenSubscribeTerminator.cs" />
-    <Compile Include="Routing\MessageDrivenSubscriptions\MessageDrivenSubscriptions.cs" />
     <Compile Include="Routing\DistributionStrategy.cs" />
     <Compile Include="Routing\MessagingBestPractices\EnforceBestPracticesOptions.cs" />
     <Compile Include="Routing\MessagingBestPractices\EnforcePublishBestPracticesBehavior.cs" />
