+++ src/NServiceBus.Core/Persistence/InMemory/SubscriptionStorage/InMemorySubscriptionPersistence.cs
@@ -7,7 +7,6 @@
     {
         internal InMemorySubscriptionPersistence()
         {
-            DependsOn<MessageDrivenSubscriptions>();
         }
 
         /// <summary>
