+++ src/NServiceBus.Core/obsoletes.cs
@@ -2603,6 +2603,17 @@ namespace NServiceBus.Features
         {
         }
     }
+
+    [ObsoleteEx(
+        RemoveInVersion = "7",
+        TreatAsErrorFromVersion = "7")]
+    public class MessageDrivenSubscriptions : Feature
+    {
+        protected internal override void Setup(FeatureConfigurationContext context)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
 }
